move __progname handling to its own file
